SugarBounce Community Discord Bot/requirements.txt
requests==2.27.1
discord.py==1.7.3
tweepy==4.7.0